requirements.txt
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
filelock==3.0.12
geomloss==0.2.4
h5py==3.3.0
idna==2.10
importlib-metadata==3.4.0
joblib==1.0.0
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014532738/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-base_1603376012865/work
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nltk==3.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
olefile==0.46
packaging==20.8
pandas==1.2.3
Pillow @ file:///tmp/build/80754af9/pillow_1609786740831/work
pykeops==1.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.0
sentence-transformers==0.4.1.2
sentencepiece==0.1.95
six @ file:///tmp/build/80754af9/six_1605205313296/work
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.56.0
transformers==4.2.2
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.3
zipp==3.4.0
